+++ packages/frontend/src/features/catalog/components/CatalogPanel.tsx
@@ -22,6 +22,7 @@ import { useHistory } from 'react-router-dom';
 import MantineIcon from '../../../components/common/MantineIcon';
 import { useProject } from '../../../hooks/useProject';
 import { useCatalog } from '../hooks/useCatalog';
+import { useCatalogAnalytics } from '../hooks/useCatalogAnalytics';
 import { useCatalogMetadata } from '../hooks/useCatalogMetadata';
 import { CatalogMetadata } from './CatalogMetadata';
 import { CatalogTree } from './CatalogTree';
@@ -115,6 +116,11 @@ export const CatalogPanel: FC<React.PropsWithChildren<Props>> = ({
         data: metadata,
         reset: closeMetadata,
     } = useCatalogMetadata(projectUuid);
+    const {
+        mutate: getAnalytics,
+        isLoading: isAnalyticsLoading,
+        data: analytics,
+    } = useCatalogAnalytics(projectUuid);
 
     const handleSearchChange = useCallback(
         (searchString: string) => {
@@ -187,11 +193,13 @@ export const CatalogPanel: FC<React.PropsWithChildren<Props>> = ({
                 const table = catalogResults.find(
                     (item) => item.name === selectedItem.table,
                 );
-                if (table && table.type === CatalogType.Table)
+                if (table && table.type === CatalogType.Table) {
                     getMetadata(selectedItem.table);
+                    getAnalytics(selectedItem.table);
+                }
             }
         },
-        [catalogResults, getMetadata],
+        [catalogResults, getMetadata, getAnalytics],
     );
 
     const history = useHistory();
@@ -355,9 +363,11 @@ export const CatalogPanel: FC<React.PropsWithChildren<Props>> = ({
 
                     {metadata && (
                         <CatalogMetadata
-                            data={metadata}
+                            metadataResults={metadata}
                             projectUuid={projectUuid}
                             selection={selection}
+                            analyticResults={analytics}
+                            isAnalyticsLoading={isAnalyticsLoading}
                         />
                     )}
                 </Stack>